requirements.txt
streamlit>=1.37.0
pydantic>=2.7.0
python-dotenv>=1.0.1
pypdf>=4.2.0
openai>=1.40.0
tiktoken>=0.7.0
tenacity>=8.3.0
